[crypto] Rename lib/crypto/impl/ecdsa_p256 to lib/crypto/impl/ecdsa.
This is a minor fixup that will ultimately result in slightly less
clutter in the impl/ directory, allowing ecdsa/ to contain all ECDSA
implementations for different curves rather than separating different
curves into different directories.
