--- sw/device/lib/crypto/impl/ecdsa_p256/README.md
@@ -1,4 +1,0 @@
-# ECDSA with NIST P-256
-
-ECDSA implementation with the P-256 curve, following [FIPS
-186-4](https://nvlpubs.nist.gov/nistpubs/FIPS/NIST.FIPS.186-4.pdf).
